Removed deprecated for 2.13
