+++ src/main/scala-2.12/reactor/core/scala/publisher/VersionedSFluxCompanion.scala
@@ -1,0 +1,8 @@
+package reactor.core.scala.publisher
+
+import reactor.core.publisher.{Flux => JFlux}
+import reactor.core.scala.publisher.versioned._
+
+trait VersionedSFluxCompanion {
+  def fromStream[T](streamSupplier: () => Stream[T]): SFlux[T] = new ReactiveSFlux[T](JFlux.fromStream[T](streamSupplier()))
+}
