+++ src/main/scala-2.13/reactor/core/scala/publisher/VersionedFluxProcessor.scala
@@ -1,0 +1,9 @@
+package reactor.core.scala.publisher
+
+import reactor.core.scala.{Scannable, VersionedScannable}
+
+import scala.jdk.CollectionConverters._
+
+trait VersionedFluxProcessor[IN, OUT] extends VersionedScannable { self: FluxProcessor[IN, OUT] =>
+  override def inners(): LazyList[_ <: Scannable] = jFluxProcessor.inners().iterator().asScala.map(js=> js: Scannable).to(LazyList)
+}
