+++ src/main/scala-2.13/reactor/core/scala/publisher/VersionedSFluxCompanion.scala
@@ -1,0 +1,8 @@
+package reactor.core.scala.publisher
+
+import reactor.core.publisher.{Flux => JFlux}
+import scala.jdk.StreamConverters._
+
+trait VersionedSFluxCompanion {
+  def fromStream[T](streamSupplier: () => LazyList[T]): SFlux[T] = new ReactiveSFlux[T](JFlux.fromStream[T](streamSupplier().asJavaSeqStream))
+}
