+++ src/main/scala-2.13/reactor/core/scala/publisher/versioned/package.scala
@@ -1,0 +1,10 @@
+package reactor.core.scala.publisher
+
+import java.util.stream.{Stream => JStream}
+
+import scala.language.implicitConversions
+import scala.jdk.StreamConverters._
+
+package object versioned {
+  implicit def scalaStream2JavaStream[T](stream: LazyList[T]): JStream[T] = stream.asJavaSeqStream
+}
