+++ src/test/scala/reactor/core/scala/publisher/SFluxTest.scala
@@ -166,14 +166,6 @@ class SFluxTest extends AnyFreeSpec with Matchers with TableDrivenPropertyChecks
         .verifyComplete()
     }
 
-    ".fromStream" - {
-      "with supplier should create flux that emit items contained in the supplier" in {
-        StepVerifier.create(SFlux.fromStream(() => Stream(1, 2, 3)))
-          .expectNext(1, 2, 3)
-          .verifyComplete()
-      }
-    }
-
     ".generate" - {
       "with state supplier and state consumer" in {
         val tempFile = Files.createTempFile("fluxtest-", ".tmp").toFile
@@ -437,41 +429,6 @@ class SFluxTest extends AnyFreeSpec with Matchers with TableDrivenPropertyChecks
         .verifyComplete()
     }
 
-    ".using" - {
-      "without eager flag should produce some data" in {
-        val tempFile = Files.createTempFile("fluxtest-", ".tmp")
-        tempFile.toFile.deleteOnExit()
-        new PrintWriter(tempFile.toFile) {
-          write(s"1${sys.props("line.separator")}2")
-          flush()
-          close()
-        }
-
-        StepVerifier.create(
-          SFlux.using[String, File](() => tempFile.toFile, (file: File) => SFlux.fromIterable[String](Source.fromFile(file).getLines().toIterable), (file: File) => {
-            file.delete()
-            ()
-          }))
-          .expectNext("1", "2")
-          .verifyComplete()
-      }
-      "with eager flag should produce some data" in {
-        val tempFile = Files.createTempFile("fluxtest-", ".tmp")
-        tempFile.toFile.deleteOnExit()
-        new PrintWriter(tempFile.toFile) {
-          write(s"1${sys.props("line.separator")}2")
-          flush()
-          close()
-        }
-        StepVerifier.create(
-          SFlux.using[String, File](() => tempFile.toFile, (file: File) => SFlux.fromIterable[String](Source.fromFile(file).getLines().toIterable), (file: File) => {
-            file.delete()
-            ()
-          }, eager = true))
-          .expectNext("1", "2")
-          .verifyComplete()
-      }
-    }
 
     ".zip" - {
       "with source1, source2 and combinator should combine the data" in {
@@ -585,26 +542,6 @@ class SFluxTest extends AnyFreeSpec with Matchers with TableDrivenPropertyChecks
     }
 
     ".bracketCase" - {
-      "should release all resources properly" in {
-        import java.io.PrintWriter
-        val files = (0 until 5) map(i => {
-          val path = Files.createTempFile(s"bracketCase-$i", ".tmp")
-          val file = path.toFile
-          new PrintWriter(file) { write(s"$i"); close() }
-          file
-        })
-        files.foreach(f => f.exists() shouldBe true)
-        val sf = SFlux.fromIterable(files)
-          .bracketCase(f => {
-            SFlux.just(Source.fromFile(f))
-              .bracket(br => SFlux.fromIterable(br.getLines().toIterable))(_.close())
-          })((file, _) => file.delete())
-        StepVerifier.create(sf)
-          .expectNext("0", "1", "2", "3", "4")
-          .verifyComplete()
-        files.foreach(f => f.exists() shouldBe false)
-      }
-
       "should handle ExitCase.error" - {
         "when the error happens inside the generated flux" in {
           import java.io.PrintWriter
@@ -939,16 +876,6 @@ class SFluxTest extends AnyFreeSpec with Matchers with TableDrivenPropertyChecks
       "with keyExtractor and valueExtractor should collect the value, extract the key and value from it" in {
         StepVerifier.create(SFlux.just(1, 2, 3, 4, 5, 6, 7, 8, 9, 10).collectMultimap(i => i % 3, i => i + 6))
           .expectNext(Map((0, Seq(9, 12, 15)), (1, Seq(7, 10, 13, 16)), (2, Seq(8, 11, 14))))
-          .verifyComplete()
-      }
-      "with keyExtractor, valueExtractor and map supplier should collect the value, extract the key and value from it and put in the provided map" in {
-        val map = mutable.HashMap[Int, util.Collection[Int]]()
-        StepVerifier.create(SFlux.just(1, 2, 3, 4, 5, 6, 7, 8, 9, 10).collectMultimap(i => i % 3, i => i + 6, () => map))
-          .expectNextMatches((m: Map[Int, Traversable[Int]]) => {
-            m shouldBe map.mapValues(vs => vs.toArray().toSeq).toMap
-            m shouldBe Map((0, Seq(9, 12, 15)), (1, Seq(7, 10, 13, 16)), (2, Seq(8, 11, 14)))
-            true
-          })
           .verifyComplete()
       }
     }
@@ -2130,11 +2057,6 @@ class SFluxTest extends AnyFreeSpec with Matchers with TableDrivenPropertyChecks
         .verifyComplete()
     }
 
-    ".tag should tag the Flux and accessible from Scannable" in {
-      val flux = SFlux.just(1, 2, 3).tag("integer", "one, two, three")
-      Scannable.from(Option(flux)).tags shouldBe Stream("integer" -> "one, two, three")
-    }
-
     ".tail should return flux that exclude the head" in {
       StepVerifier.create(SFlux.just(1, 2, 3, 4, 5).tail)
         .expectNext(2, 3, 4, 5)
@@ -2252,15 +2174,6 @@ class SFluxTest extends AnyFreeSpec with Matchers with TableDrivenPropertyChecks
         val list = SFlux.just(1, 2, 3)
           .toIterable[Int](1, Option(Queues.get(1))).toList
         list shouldBe Iterable(1, 2, 3)
-      }
-    }
-
-    ".toStream" - {
-      "should transform this flux into stream" in {
-        SFlux.just(1, 2, 3).toStream() shouldBe Stream(1, 2, 3)
-      }
-      "with batchSize should transform this flux into stream" in {
-        SFlux.just(1, 2, 3).toStream(2) shouldBe Stream(1, 2, 3)
       }
     }
